changes made out

+++ SMD.Web/Areas/Coupons/Scripts/Layout.viewModel.js
@@ -1,5 +1,5 @@
 ﻿define("Layout/Layout.viewModel",
-    ["jquery", "amplify", "ko", "Layout/Layout.dataService", "Layout/Layout.model", "common/confirmation.viewModel"],
+    ["jquery", "amplify", "ko", "Layout/Layout.dataService", "Layout/Layout.model","common/confirmation.viewModel"],
 
     function ($, amplify, ko, dataService, model, confirmation) {
 
@@ -15,6 +15,7 @@
                 selectedCategory = ko.observable(),
                 selectedBranchField = ko.observable(),
                 isSaveChangesEnable = ko.observable(false),
+                isMapVisible = ko.observable(false);
                 isdeleteEnable = ko.observable(true),
                 counter = ko.observable(0),
                 defaultOpenBranchFieldName = ko.observable(),
@@ -56,6 +57,8 @@
                                 selectedBranch(branch);
                                 isSaveChangesEnable(true);
                                 isdeleteEnable(true);
+                                codeAddress();
+                                isMapVisible(true);
 
                                 //}
 
@@ -80,6 +83,7 @@
                                                                   item.brachFeilds.push(selectedBranch());
                                                                   isSaveChangesEnable(false);
                                                                   isdeleteEnable(false);
+                                                                  isMapVisible(false);
                                                                   toastr.success("Successfully saved.");
 
                                                               }
@@ -95,6 +99,7 @@
                                                                   selectedCategory().isExpanded(false);
                                                                   isSaveChangesEnable(false);
                                                                   isdeleteEnable(false);
+                                                                  isMapVisible(false);
 
                                                               }
                                                           })
@@ -217,9 +222,23 @@
 
                 },
                 DeleteCategory = function (category) {
-                    confirmation.messageText("Please delete or move all branches from this category to delete it.");
+                    if (category.brachFeilds().length > 0) {
+                        confirmation.showOKpopup();
+
+                    }
+                    else {
+
+                        confirmation.messageText("Do you want to delete Branch?");
+                        confirmation.show();
+
+                    }
                     confirmation.afterProceed(function () {
-                        selectedCategory(category);
+                        if (category.brachFeilds().length > 0)
+                        {
+                            return;
+                        }
+                        else {
+                            selectedCategory(category);
                         dataService.DeleteCurrentCategory(
                                     category.convertToServerData(), {
                                         success: function (data) {
@@ -230,12 +249,14 @@
                                             toastr.error("Failed to Delete Category . Error: " + response);
                                         }
                                     });
+                        }
+                        
                     });
 
                     confirmation.afterCancel(function () {
                         confirmation.hide();
                     });
-                    confirmation.show();
+                   
                 },
                 EditCategory = function (category) {
                     selectedCategory(category)
@@ -250,6 +271,20 @@
 
 
                 },
+                  isAddressFilled = ko.computed(function () {
+                      if (selectedBranch() != undefined && selectedBranch() != null) {
+                          if ((selectedBranch().branchAddressline1() == undefined || selectedBranch().branchAddressline1() == "") || (selectedBranch().branchCity() == undefined || selectedBranch().branchCity() == "") || (selectedBranch().branchState() == undefined || selectedBranch().branchState() == "") ||(selectedBranch().branchZipCode() == undefined || selectedBranch().branchZipCode() == "")) {
+                              return false;
+                          }
+                          return true;
+                      }
+                      else {
+
+                          return false;
+                      }
+                     
+                      
+                  }),
                 CreateNewBranchLocation = function () {
 
                     var newBranchLocation = model.Branch({});
@@ -257,6 +292,7 @@
                     selectedBranch(newBranchLocation);
                     isSaveChangesEnable(true);
                     isdeleteEnable(false);
+                    isMapVisible(true);
                 },
                 resetTreeExpensionAfterSave = function (category) {
                     category.isExpanded(true);
@@ -271,6 +307,7 @@
                         selectedBranch(null);
                         isSaveChangesEnable(false);
                         isdeleteEnable(false);
+                        isMapVisible(false);
                     }
 
                     else if (selectedBranch() != null && selectedBranch() != undefined && selectedBranch().hasChanges()) {
@@ -283,6 +320,7 @@
                         selectedBranch(null);
                         isSaveChangesEnable(false);
                         isdeleteEnable(false);
+                        isMapVisible(false);
 
                     }
                     confirmation.afterCancel(function () {
@@ -290,6 +328,7 @@
                         selectedBranch(null);
                         isSaveChangesEnable(false);
                         isdeleteEnable(false);
+                        isMapVisible(false);
                         confirmation.hide();
 
 
@@ -336,6 +375,7 @@
                     // branches.removeAll();
                     selectedBranchField(brachFeilds);
                     getBranchesByBranchFieldId(brachFeilds.branchId());
+                  
 
                 },
                 selectCategory = function (category, event) {
@@ -360,7 +400,6 @@
                     codeAddress();
                     google.maps.event.addDomListener(window, 'load', initializeGEO);
                 },
-
                initializeGEO = function () {
                    geocoder = new google.maps.Geocoder();
                    var latlng = new google.maps.LatLng(-34.397, 150.644);
@@ -373,7 +412,7 @@
 
                }
                 codeAddress = function () {
-                   // var address = selectedBranch().branchAddressline1().toLowerCase() + ',' + selectedBranch().branchCity().toLowerCase() + ',' + selectedBranch().branchZipCode() + ',' + selectedBranch().branchState().toLowerCase();
+                    // var address = selectedBranch().branchAddressline1().toLowerCase() + ',' + selectedBranch().branchCity().toLowerCase() + ',' + selectedBranch().branchZipCode() + ',' + selectedBranch().branchState().toLowerCase();
                     var address = selectedBranch().branchAddressline1().toLowerCase() + ' ' + selectedBranch().branchCity().toLowerCase() + ' ' + selectedBranch().branchZipCode() + ' ' + selectedBranch().branchState().toLowerCase();
                     geocoder.geocode({
                         'address': address
@@ -381,14 +420,15 @@
                         if (status == google.maps.GeocoderStatus.OK) {
                             selectedBranch().branchLocationLat(results[0].geometry.location.lat());
                             selectedBranch().branchLocationLon(results[0].geometry.location.lng());
-                           map.setCenter(results[0].geometry.location);
-      
+                            map.setCenter(results[0].geometry.location);
+
                             var marker = new google.maps.Marker({
                                 map: map,
                                 position: results[0].geometry.location
                             });
                         } else {
-                            alert('Geocode was not successful for the following reason: ' + status);
+                            toastr.error("Failed to Search Address,please add valid address and search it . Error: " + status);
+                            //alert('Geocode was not successful for the following reason: ' + status);
                         }
                     });
                 }
@@ -453,8 +493,10 @@
                     EditCategory: EditCategory,
                     SaveCategory: SaveCategory,
                     DeleteCategory: DeleteCategory,
+                    isMapVisible: isMapVisible,
                     isSaveChangesEnable: isSaveChangesEnable,
                     isdeleteEnable: isdeleteEnable,
+                    isAddressFilled: isAddressFilled,
                     hideBranchCategoryDialog: hideBranchCategoryDialog,
                     CodeAddressonMap: CodeAddressonMap,
                     branchDdlist: branchDdlist
